Slate: Fix some implicit conversion in SlateCore code. #jira UE-88354
[CL 14601543 by Patrick Boutot in ue5-main branch]

+++ Engine/Source/Runtime/SlateCore/Public/Animation/CurveSequence.h
@@ -210,7 +210,7 @@ private:
 	double PauseTime;
 
 	/** How long the entire sequence lasts. */
-	float TotalDuration;
+	double TotalDuration;
 
 	/** Are we playing the animation in reverse */
 	uint8 bInReverse : 1;
